Version 1.00: Debug the op.py file class Linear
Debug the op.py file class Linear forward and backward self.W to self.params['W'] and self.b to self.params['b'] to fix the bug in the forward and backward methods.
Use params to access the weights and biases in the Linear class prevents potential issues of uncoupling the params from the class.